Move styles from .css file to <style> tag
Though this flies in the face of standard practice, it also enables the
resume to be distributed as a standalone html file with all* of its
styling intact.

This also motivates the addition of the .gitignore file. I created a
second (hard) link to the html file on disk, so I could use the file as
both an easily-served `index.html` and as a nicely-named
`alex_birdsall_resume.html` to distribute without having to go to the
bother of manually copying shit.

\* Okay, the webfonts are still loaded from a googlefonts url, but there
are reasonable fallbacks in place for offline viewing.