Remove unused packages, remove yarn
With the gatsby build, I was stuck with this weird chimera project where
I had to use npm to install the dependencies without error and then use
yarn to run the project without error.

This has many fewer dependencies, is much easier to upgrade if I choose,
and actually works with a fresh install via a single package manager.
What a concept!